+++ src/main/java/hexlet/code/service/UserService.java
@@ -4,9 +4,11 @@ package hexlet.code.service;
 import hexlet.code.dto.UserCreateDTO;
 import hexlet.code.dto.UserDTO;
 import hexlet.code.dto.UserUpdateDTO;
+import hexlet.code.exception.MethodNotAllowedException;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.UserMapper;
 import hexlet.code.model.User;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -25,6 +27,9 @@ public class UserService {
 
     @Autowired
     private PasswordEncoder encoder;
+
+    @Autowired
+    private TaskRepository taskRepository;
 
     public List<UserDTO> getAll() {
         List<User> users = userRepository.findAll();
@@ -62,6 +67,9 @@ public class UserService {
     }
 
     public void delete(Long id) {
+        if (!taskRepository.findByAssigneeId(id).isEmpty()) {
+            throw new MethodNotAllowedException("You cannot delete a user. The user is associated with tasks.");
+        }
         userRepository.deleteById(id);
     }
 }
